backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "theAgency — AI Development Pipeline Orchestrator backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.4",
    "fastapi>=0.133.1",
    "greenlet>=3.3.2",
    "httpx>=0.28.1",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.47",
    "sse-starlette>=3.2.0",
    "structlog>=25.5.0",
    "uvicorn[standard]>=0.41.0",
    "watchfiles>=1.1.1",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "respx>=0.22.0",
    "ruff>=0.15.4",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true
